requirements.txt
matplotlib==3.4.3
Pillow==8.3.2
tensorflow==2.6.0
